mmr sorting
