+++ civiclens/utils/constants.py
@@ -1,5 +1,7 @@
 import os
+
 from dotenv import load_dotenv
+
 
 load_dotenv(override=True)  # intentionally override any existing set env vars
