+++ pyproject.toml
@@ -23,6 +23,7 @@ nltk = "^3.8.1"
 bertopic = "^0.16.1"
 spacy = "^3.7.4"
 pandas = "^2.2.2"
+langchain = "^0.1.17"
 
 [tool.poetry.group.dev.dependencies]
 ruff = "^0.3.5"
